+++ website/src/examples/basic_buttons.rs
@@ -1,5 +1,5 @@
 use leptos::*;
-use leptos_bulma::elements::{BButton, BButtons};
+use leptos_bulma::elements::{BAButton, BButton, BButtons};
 use leptos_bulma::enums::BColor;
 
 #[component]
@@ -7,6 +7,9 @@ pub fn BasicButtons() -> impl IntoView {
     view! {
         <BButtons>
             <BButton>"Normal button"</BButton>
+            <BAButton href="/elements#button" target="_blank" title="Open in a new tab">
+                "Anchor link button"
+            </BAButton>
             <BButton is_light=true>"Light button"</BButton>
             <BButton is_dark=true>"Dark button"</BButton>
             <BButton is_inverted=true>"Inverted button"</BButton>